Basic TRX Viewer in QT
I couldn't find a way to show all of the tests in HTML without a weird
client-server relationship which I didn't want to mimic because it
wouldn't be comfortable.

So I created a QT viewer instead of HTML viewer. Besides the TRX
handling all of the logic is different, so it's not that related to the
rest of the tree.